Исправляет ошибки в ДЗ
