+++ index.html
@@ -105,9 +105,7 @@
 
                 <!-- Добавление комментария к изображению -->
                 <fieldset class="img-upload__text text">
-                    <textarea class="text__description" name="description" placeholder="Ваш комментарий..." required minlength="20" maxlength="140" data-pristine-required-message="Обязательное поле"
-                    data-pristine-minlength-message="Минимальная длина 20 символов"
-                    data-pristine-maxlength-message="Максимальная длина 140 символов"></textarea>
+                    <textarea class="text__description" name="description" placeholder="Ваш комментарий..." required minlength="20" maxlength="140"></textarea>
                 </fieldset>
 
                 <!-- Кнопка для отправки данных на сервер -->
@@ -166,8 +164,6 @@
       </div>
     </section>
   </template>
-
-<script src="vendor/pristine/pristine.min.js" type="text/javascript"></script>
 <script src="js/main.js" type="module"></script>
 </body>
 </html>
