+++ js/form.js
@@ -1,0 +1,40 @@
+import { isEscapeKey } from './util.js';
+
+const uploadFileInput = document.querySelector('#upload-file');
+const imageUploadForm = document.querySelector('.img-upload__overlay');
+const body = document.querySelector('body');
+const cancelButton = document.querySelector('#upload-cancel');
+function onModalEscKeydown (evt) {
+  if (isEscapeKey(evt)) {
+    evt.preventDefault();
+    closeModal();
+  }
+}
+function onCancelButtonClick () {
+  closeModal();
+}
+
+// открывает окно редактирования фото
+function openModal () {
+  imageUploadForm.classList.remove('hidden');
+  body.classList.add('.modal-open');
+  document.addEventListener('keydown', onModalEscKeydown);
+  cancelButton.addEventListener('click', onCancelButtonClick);
+}
+
+// закрывает окно редактирования фото
+function closeModal () {
+  imageUploadForm.classList.add('hidden');
+  body.classList.remove('.modal-open');
+  document.removeEventListener('keydown', onModalEscKeydown);
+  cancelButton.removeEventListener('click', onCancelButtonClick);
+  uploadFileInput.value = '';
+}
+
+function initForm () {
+  uploadFileInput.addEventListener('change', () => {
+    openModal();
+  });
+}
+
+export {initForm};
