+++ js/main.js
@@ -1,6 +1,6 @@
 import {getPhotos} from './get-photos.js'; // модуль с функцией для генерации фотокарточек
 import {renderPhotos} from './rendering.js';
-import { initForm } from './init-form.js';
+import { initForm } from './form.js';
 
 const photos = getPhotos();
 renderPhotos(photos);
